+++ aws_s3/utils/aws_s3_webtx_validator.py
@@ -28,24 +28,31 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+AWS S3 WebTx validator.
 """
 
-"""AWS S3 WebTx validator."""
-
-
 import boto3
+from botocore.exceptions import NoCredentialsError
 from botocore.config import Config
 from .aws_s3_webtx_constants import REGIONS
+from .aws_s3_exceptions import AWSS3WebTXException
 
 
 class AWSS3WebTxValidator(object):
     """AWS S3 WebTx validator class."""
 
-    def __init__(self, logger, proxy):
+    def __init__(self, configuration, logger, proxy, storage, log_prefix):
         """Initialize."""
         super().__init__()
+        self.configuration = configuration
         self.logger = logger
         self.proxy = proxy
+        self.storage = storage
+        self.log_prefix = log_prefix
+        self.aws_public_key = None
+        self.aws_private_key = None
+        self.aws_session_token = None
 
     def validate_max_file_size(self, max_file_size):
         """Validate max file size.
@@ -54,7 +61,8 @@ class AWSS3WebTxValidator(object):
             max_file_size: the max file size to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         if max_file_size:
             try:
@@ -74,7 +82,8 @@ class AWSS3WebTxValidator(object):
             max_duration: the max duration to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         if max_duration:
             try:
@@ -94,7 +103,8 @@ class AWSS3WebTxValidator(object):
             region_name: the region name to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         if region_name:
             try:
@@ -108,26 +118,51 @@ class AWSS3WebTxValidator(object):
         else:
             return False
 
-    def validate_credentials(self, aws_public_key, aws_private_key):
+    def validate_credentials(self, aws_client):
         """Validate credentials.
 
         Args:
-            aws_public_key: the aws public key to establish connection with aws s3.
-            aws_private_key: the aws private key to establish connection with aws s3.
+            aws_public_key: the aws public key to establish connection
+            with AWS S3.
+            aws_private_key: the aws private key to establish connection
+            with AWS S3.
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not. True in case of
+            valid value, False otherwise
         """
         try:
             s3_resource = boto3.resource(
                 "s3",
-                aws_access_key_id=aws_public_key,
-                aws_secret_access_key=aws_private_key,
+                aws_access_key_id=aws_client.aws_public_key,
+                aws_secret_access_key=aws_client.aws_private_key,
+                aws_session_token=aws_client.aws_session_token,
+                region_name=self.configuration.get("region_name"),
                 config=Config(proxies=self.proxy),
             )
+
             for _ in s3_resource.buckets.all():
                 break
             return True
-        except Exception as e:
-            self.logger.error(f"AWS S3 WebTx Plugin: Error while validating creadentials: {e}") 
-            raise
+        except NoCredentialsError as exp:
+            err_msg = (
+                "No AWS Credentials were found in the environment."
+                " Deploy the plugin into AWS environment or use AWS IAM "
+                "Roles Anywhere authentication method."
+            )
+            self.logger.error(
+                message=f"{self.log_prefix}: {err_msg}.",
+                details=f"Error: {exp}",
+            )
+            raise AWSS3WebTXException(err_msg)
+        except Exception as exp:
+            err_msg = (
+                "Error occurred while validating credentials. "
+                "Make sure all the required bucket permissions "
+                "are attached to the user."
+            )
+            self.logger.error(
+                message=f"{self.log_prefix}: {err_msg}.",
+                details=f"Error: {exp}",
+            )
+            raise AWSS3WebTXException(err_msg)