+++ aws_s3/CHANGELOG.md
@@ -1,5 +1,13 @@
+# 1.2.0
+## Added
+- Added two new authentication methods.
+## Removed
+- Removed secret credentials for authentication.
+## Changed
+- Updated object structure in push functionality.
+
 # 1.1.0
-## Modified
+## Changed
 - Added support for few regions.
 - Changed the plugin name